[FIX] moove the main part of the chatbot to the component directory
