+++ crypto-pilot-builder/.gitignore
@@ -28,3 +28,5 @@ coverage
 *.sw?
 
 *.tsbuildinfo
+
+__pycache__
